refactor(Main, WiseSaying): Main 파일에서 클래스 분리
- WiseSaying 클래스로 분리
