+++ src/main/java/ex/Main.java
@@ -1,38 +1,8 @@
-package org.example;
+package ex;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Scanner;
-
-class WiseSaying {
-    private int number;
-    private String author;
-    private String content;
-
-    public WiseSaying(int number, String author, String content) {
-        this.number = number;
-        this.author = author;
-        this.content = content;
-    }
-
-    public void up(){
-        System.out.println(number+" / " +author+" / "+content);
-    }
-
-    public int getNumber(){
-        return number;
-    }
-
-    public void editWiseSaying(){
-        System.out.println("명언(기존) : " + content);
-        System.out.print("명언 : ");
-        content = new Scanner(System.in).nextLine();
-
-        System.out.println("작가(기존) : " + author);
-        System.out.print("작가 : ");
-        author = new Scanner(System.in).nextLine();
-    }
-}
 
 public class Main {
     public static void main(String[] args) {
